Java Project that finds list of scrabble words from given userInput.
“You are given a dictionary (dictionary.txt), containing a list of
words, one per line. Imagine you have seven tiles. Each tile is either
blank or contains a single lowercase letter (a-z).

Please list all the words from the dictionary that can be produced by
using some or all of the seven tiles, in any order. A blank tile is a
wildcard, and can be used in place of any letter.

Try to use a minimal amount of memory.

1. Find all of the words that can be formed if you don't have to deal
with blank tiles.

2. Find all of the words that can be formed, including those where
blank tiles are used as wildcards.

3. Please bear in mind you will need to process several hundred of
7-tile sets with the same dictionary.
 